requirement.txt
Flask==2.0.2
Werkzeug==2.0.2
opencv-python==4.5.4.58
numpy==1.21.4
Pillow==8.4.0
pdf2image==1.16.0
imutils==0.5.4